additional check not to delete if razvozka fulfilled

+++ razv4_0/views/rzv_attributes.py
@@ -27,5 +27,6 @@ def razvozka_deliver_to(request, razv_id):
 
 def razvozka_delete(request, razv_id):
     razv_to_delete = Razvozka.objects.get(id=razv_id)
-    razv_to_delete.delete()
+    if not razv_to_delete.fulfilled:
+        razv_to_delete.delete()
     return HttpResponse()
